2.realtime-data-ingestion/requirements.txt
google-cloud-pubsub
apache-beam[gcp]